+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 - Confirmation when trying to close the page with unsaved changes to the poem
 - PokeHelp mode: highlighting rhymes, syllable counter, text counters (words, characters, lines)
+- PokeMarkov
 
 ## Fixed
 - Broken notebook panes when resizing